rename: performaceHandlers를 performancesHandlers로 병합
