+++ src/mocks/server.ts
@@ -1,5 +1,4 @@
 import { setupServer } from 'msw/node';
-import { performancesHandlers } from './handlers/performanceHandlers';
 import { performancesHandlers } from './handlers/performancesHandlers';
 import { reviewsHandlers } from './handlers/reviewsHandlers';
